+++ test/lightning_web/controllers/api/provisioning_controller_test.exs
@@ -211,8 +211,7 @@ defmodule LightningWeb.API.ProvisioningControllerTest do
           name: "workflow 1"
         )
 
-      {:ok, snapshot_1} =
-        Snapshot.get_or_create_latest_for(workflow_1, insert(:user))
+      {:ok, snapshot_1} = Snapshot.create(workflow_1)
 
       {:ok, updated_workflow_1} =
         workflow_1
@@ -225,8 +224,7 @@ defmodule LightningWeb.API.ProvisioningControllerTest do
           name: "workflow 2"
         )
 
-      {:ok, snapshot_2} =
-        Snapshot.get_or_create_latest_for(workflow_2, insert(:user))
+      {:ok, snapshot_2} = Snapshot.create(workflow_2)
 
       conn =
         get(conn, ~p"/api/provision/#{project_id}", snapshots: [snapshot_1.id])
@@ -428,8 +426,7 @@ defmodule LightningWeb.API.ProvisioningControllerTest do
           name: "workflow 1"
         )
 
-      {:ok, snapshot_1} =
-        Snapshot.get_or_create_latest_for(workflow_1, insert(:user))
+      {:ok, snapshot_1} = Snapshot.create(workflow_1)
 
       {:ok, updated_workflow_1} =
         workflow_1
@@ -442,8 +439,7 @@ defmodule LightningWeb.API.ProvisioningControllerTest do
           name: "workflow 2"
         )
 
-      {:ok, snapshot_2} =
-        Snapshot.get_or_create_latest_for(workflow_2, insert(:user))
+      {:ok, snapshot_2} = Snapshot.create(workflow_2)
 
       conn =
         Plug.Conn.put_req_header(